+++ meta/def.rb
@@ -2,6 +2,7 @@ Mutant::Meta::Example.add do
   source 'def foo; end'
 
   mutation 'def foo; raise; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -11,6 +12,7 @@ Mutant::Meta::Example.add do
   mutation 'def foo; nil; rescue; end'
   mutation 'def foo; self; rescue; end'
   mutation 'def foo; end'
+  mutation 'remove_method :foo'
 
   # Promote rescue resbody bodies
   mutation 'def foo; foo; end'
@@ -41,6 +43,8 @@ Mutant::Meta::Example.add do
 
   # Failing body
   mutation 'def a; raise; end'
+
+  mutation 'remove_method :a'
 end
 
 Mutant::Meta::Example.add do
@@ -60,6 +64,8 @@ Mutant::Meta::Example.add do
   mutation 'def foo; end'
 
   mutation 'def foo; raise; end'
+
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -78,6 +84,8 @@ Mutant::Meta::Example.add do
 
   # Mutation of body
   mutation 'def foo(a, b); raise; end'
+
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -94,6 +102,7 @@ Mutant::Meta::Example.add do
   mutation 'def foo(b = nil); true; end'
   mutation 'def foo(a, _b = nil); true; end'
   mutation 'def foo(a, b); true; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -101,6 +110,7 @@ Mutant::Meta::Example.add do
 
   mutation 'def foo(_unused); raise; end'
   mutation 'def foo; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -111,6 +121,7 @@ Mutant::Meta::Example.add do
   mutation 'def foo(_unused = true); raise; end'
   mutation 'def foo(_unused); end'
   mutation 'def foo; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -132,6 +143,7 @@ Mutant::Meta::Example.add do
   mutation 'def foo(a = 0, b = 0); a = 0; end'
   mutation 'def foo(a = 0, b = 0); b = 0; end'
   mutation 'def foo(a = 0, b = 0); raise; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do
@@ -144,6 +156,7 @@ Mutant::Meta::Example.add do
   mutation 'def foo(_a = true); end'
   mutation 'def foo(a = true); raise; end'
   mutation 'def foo(a = true); a = true; end'
+  mutation 'remove_method :foo'
 end
 
 Mutant::Meta::Example.add do